+++ bin/AssemblePairs.py
@@ -8,7 +8,6 @@ from presto import __version__, __date__
 
 # Imports
 import os
-import tempfile
 import shutil
 import sys
 import numpy as np
@@ -16,7 +15,6 @@ import pandas as pd
 import scipy.stats as stats
 from argparse import ArgumentParser
 from collections import OrderedDict
-from io import StringIO
 from textwrap import dedent
 from time import time
 from Bio import SeqIO